Fix JGit compatibility issue about reflog reading from reftable.
Remove trailing '\n' on writing and add it on reading for reflog message in reftable backend implementation.

The reftable format was originally designed by Shawn Pearce for use in JGit. When use `git reflog` to get reflogs created by JGit, all logs will be displayed in one line. Because JGit doesn't write a trailing '\n' to reflogs. We should keep this consistent.

Signed-off-by: youtirsin <youtirsin@qq.com>